+++ src/applications/trafficsource/AS6802/TTTrafficSourceApp.cc
@@ -34,7 +34,7 @@ void TTTrafficSourceApp::initialize()
         Scheduled::initialize();
 
         SchedulerActionTimeEvent *event = new SchedulerActionTimeEvent("API Scheduler Task Event", ACTION_TIME_EVENT);
-        event->setAction_time(par("action_time").doubleValue()/findModuleWhereverInNode("oscillator",getParentModule())->par("tick").doubleValue());
+        event->setAction_time((uint32_t)(par("action_time").doubleValue()/findModuleWhereverInNode("oscillator",getParentModule())->par("tick").doubleValue()));
         event->setDestinationGate(gate("schedulerIn"));
         period->registerEvent(event);
     }
